Rename SmBios.h to Smbios.h in MdePkg/Include
Rename SmBios.h to Smbios.h in MdePkg/Include/IndustryStandard and
MdePkg/Include/Guid, and update references in all other places through
the edk2 tree.
