+++ EmulatorPkg/Include/Library/SmbiosLib.h
@@ -12,7 +12,7 @@ SPDX-License-Identifier: BSD-2-Clause-Patent
 #ifndef _SMBIOS_LIB_H__
 #define _SMBIOS_LIB_H__
 
-#include <IndustryStandard/SmBios.h>
+#include <IndustryStandard/Smbios.h>
 #include <Protocol/Smbios.h>
