+++ EmulatorPkg/PlatformSmbiosDxe/PlatformSmbiosDxe.c
@@ -8,7 +8,7 @@
 **/
 
 #include <PiDxe.h>
-#include <IndustryStandard/SmBios.h>
+#include <IndustryStandard/Smbios.h>
 #include <Protocol/Smbios.h>
 
 #include <Library/BaseLib.h>
