+++ MdePkg/Include/Guid/Smbios.h
