+++ OvmfPkg/EnrollDefaultKeys/EnrollDefaultKeys.c
@@ -10,7 +10,7 @@
 #include <Guid/ImageAuthentication.h>            // EFI_IMAGE_SECURITY_DATABASE
 #include <Guid/MicrosoftVendor.h>                // gMicrosoftVendorGuid
 #include <Guid/OvmfPkKek1AppPrefix.h>            // gOvmfPkKek1AppPrefixGuid
-#include <IndustryStandard/SmBios.h>             // SMBIOS_HANDLE_PI_RESERVED
+#include <IndustryStandard/Smbios.h>             // SMBIOS_HANDLE_PI_RESERVED
 #include <Library/BaseLib.h>                     // GUID_STRING_LENGTH
 #include <Library/BaseMemoryLib.h>               // CopyGuid()
 #include <Library/DebugLib.h>                    // ASSERT()
