+++ ShellPkg/Library/UefiShellDebug1CommandsLib/SmbiosView/LibSmbiosView.c
@@ -8,7 +8,7 @@
 
 
 #include "UefiShellDebug1CommandsLib.h"
-#include <Guid/SmBios.h>
+#include <Guid/Smbios.h>
 #include "LibSmbiosView.h"
 #include "SmbiosView.h"
