+++ UefiPayloadPkg/BlSupportDxe/BlSupportDxe.h
@@ -20,7 +20,7 @@ SPDX-License-Identifier: BSD-2-Clause-Patent
 #include <Library/HobLib.h>
 
 #include <Guid/Acpi.h>
-#include <Guid/SmBios.h>
+#include <Guid/Smbios.h>
 #include <Guid/SystemTableInfoGuid.h>
 #include <Guid/AcpiBoardInfoGuid.h>
 #include <Guid/GraphicsInfoHob.h>
